backend/requirements-prod.txt
# ==========================================
# Remember QRCode - Backend Dependencies
# Python 3.11.x
# ==========================================

# === Core Framework ===
fastapi==0.110.1
uvicorn[standard]==0.25.0
starlette==0.37.2

# === Database (MongoDB) ===
motor==3.3.1
pymongo==4.5.0

# === Validation ===
pydantic==2.12.5
pydantic-core==2.41.5
email-validator==2.3.0

# === Authentication (Firebase) ===
firebase-admin==7.1.0
python-jose==3.5.0
PyJWT==2.11.0

# === Payments (Mercado Pago) ===
mercadopago==2.3.0

# === Email (Resend) ===
resend==2.21.0

# === QR Code Generation ===
qrcode==8.2
pillow==12.1.0

# === HTTP Clients ===
httpx==0.28.1
requests==2.32.5
aiohttp==3.13.3

# === Utils ===
python-dotenv==1.2.1
python-multipart==0.0.22

# === Google Cloud (Firebase deps) ===
google-cloud-firestore==2.23.0
google-cloud-storage==3.9.0
google-auth==2.48.0
